+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Version - 0]
+## Fix [v0.14.1] 2022-09-22
+- fixed defect index new task
+- fixed birthDate and format in farmer & droner
 ## Feature [v0.14.0] 2022-09-22
 - edit new task
 - selected droner